--- メモ
@@ -1,70 +1,0 @@
-メモ
-⭕️困っているところ、やるべきこと
-　※一旦emailのunique外しているから注意！
-
-⭕️一番最初に/index開いた時には正常なのに、リダイレクトしたらUndefinedエラーが出てしまう。
-⭕️削除表記をしたいが、削除後にTrying to get property of non-objectと出てしまう。
-　jQueryが悪さしている感じ。
-　多分予想されるに削除されたらリダイレクトで参照仕様にもそれがないぞって感じだと思う。
-　（ない場合としての処理を用意しておく？）
-
-
-⭐️マイグレーション（完了）
-作業テーブル定義とかする。作成（up) と削除(delete) 時の設定をしておく。：
-ファイル名：migrationsフォルダ
-
-・モデル
-作業：デーブル（DB)のアクションを設定できる。
-とりあえずデータベースからとってくる処理する。特定の処理をする。
-ファイル名："ファイル名"Controller.php
-php artisan migrateがエラー起きてる。database.phpファイルを修正しないと行けなさそう。
-→envファイルとかymlファイル直したら行けた。あと純粋にマイグレーションファイルないに構文みすあり。
-＜＜Eroquent＞＞
-
-⭐️シーダー
-作業：チェック用ダミーデータの作成
-ファイル名：php artisan make:seeder CustomersTableSeeder -> CustomersTableSeeder.php
-php artisan db:seed（実行）
-※php artisan migrate:fresh --seedで実行する
-※主キー、副キーを設定した場合、副キーファイルをしっかりと作っておかないとエラーでる。
-
-・ルート
-作業：ファイルとウェブブラウザを対応させる。URLの指定と思っておく。
-ファイル名：web.php
-Route::アクション('パスの指定/index',コントローラーファイル（処理内容)）
-URLの頭に/があるかないかでだいぶ変わる。
-
-・コントローラー
-作業：全体の制御。モデルがとってきたDBファイルをViewに反映させたりする。
-コントローラー　<--> モデル　<-->　データベース
-　→コントローラーに対してルーティングする。
-ファイル名：MainController.php , CustomerController.php
-⭕️データベースから検索するものは、
-モデルの取得を使用すべき。たしかに、DBファサードを使ったりSQLコマンドのようなものを書いて
-取得することもできるが、可読性が下がる。
-
-入り口も出口も一つ。
-一つのアクションを起こしたら、一つの結果を返す。
-コントローラーの中でRequest->returnでひとまとめ。
-->コントローラーの中で処理が複数になるのは構わないが、結果一つのリターンにまとめる。
-そのために、with,compactが存在している。
-
-{{ csrf }}付けないとポストする画面で送信されない。全てのポストする画面で必要となる。
-@csrf
-
-
-・リクエスト（バリデーション）
-作業：入力値が適しているかのチェック。
-ファイル名：
-
-・ミドルウェア
-このユーザーはログインしていいのかというチェック機能
-
-
-ハマりそうなポイント
-$query = Customer::query();
-・・・if文を組み立ててクエリを組み立てる（where)を利用して。
-$customers = $query->get();
-->ここで初めてDBアクセスしてデータを持って来る。
-※入力されたら、条件に含める。
-
